Incomes Add function updated

+++ app/src/main/java/com/example/incomeexpensetracker/ui/home/HomeScreen.kt
@@ -90,7 +90,7 @@ fun homeScreen(navHostController: NavHostController) {
     }
     val accountStateList = accountViewModel.allAccounts.collectAsState()
     val accountList = accountStateList.value
-    val totalBalance = accountList.map { it.balance.toDouble() }.sumOf { it }
+    val totalBalance = accountList.map { it.balance }.sumOf { it }
     // Calculate Account  Balance>>
 
     Scaffold(
